+++ solutions/67.cpp
@@ -112,8 +112,8 @@ int main(int argc, char * argv[]) {
     UNUSED(argc);
     UNUSED(argv);
 
-    for (int i = 1; i < tri.size(); i++) {
-        for (int j = 0; j < tri[i].size(); j++) {
+    for (size_t i = 1; i < tri.size(); i++) {
+        for (size_t j = 0; j < tri[i].size(); j++) {
             if (j == 0) {
                 tri[i][j] += tri[i-1][j];
             } else if (j == tri[i].size() - 1) {
